@@ -10,32 +10,37 @@ namespace Smidge.Nuglify
 {
     public class NuglifySourceMapController : Controller
     {
-        private readonly FileSystemHelper _fileSystemHelper;
+        private readonly ISmidgeFileSystem _fileSystem;
         private readonly IBundleManager _bundleManager;
 
-        public NuglifySourceMapController(FileSystemHelper fileSystemHelper, IBundleManager bundleManager)
+        public NuglifySourceMapController(ISmidgeFileSystem fileSystem, IBundleManager bundleManager)
         {
-            _fileSystemHelper = fileSystemHelper;
+            _fileSystem = fileSystem;
             _bundleManager = bundleManager;
         }
 
         public FileResult SourceMap([FromServices] BundleRequestModel bundle)
         {
-            Bundle foundBundle;
-            if (!_bundleManager.TryGetValue(bundle.FileKey, out foundBundle))
+            if (!_bundleManager.TryGetValue(bundle.FileKey, out _))
             {
                 //TODO: Throw an exception, this will result in an exception anyways
                 return null;
             }
 
-            //now we need to determine if this bundle has already been created
-            var compositeFilePath = new FileInfo(_fileSystemHelper.GetCurrentCompositeFilePath(bundle.CacheBuster, bundle.Compression, bundle.FileKey));
-            //we need to go one level above the composite path into the non-compression named folder since the map request will always be 'none' compression
-            var mapPath = new FileInfo(Path.Combine(compositeFilePath.Directory.Parent.FullName, compositeFilePath.Name + ".map"));
-            if (mapPath.Exists)
+            var sourceMapFile = _fileSystem.CacheFileSystem.GetRequiredFileInfo(bundle.GetSourceMapFilePath());
+
+            if (sourceMapFile.Exists)
             {
-                //this should already be processed if this is being requested!
-                return File(mapPath.OpenRead(), "application/json");
+                if (!string.IsNullOrWhiteSpace(sourceMapFile.PhysicalPath))
+                {
+                    //if physical path is available then it's the physical file system, in which case we'll deliver the file with the PhysicalFileResult
+                    //FilePathResult uses IHttpSendFileFeature which is a native host option for sending static files                    
+                    return PhysicalFile(sourceMapFile.PhysicalPath, "application/json");
+                }
+                else
+                {
+                    return File(sourceMapFile.CreateReadStream(), "application/json");
+                }
             }
 
             //TODO: Throw an exception, this will result in an exception anyways
