@@ -1,11 +1,7 @@
-﻿using System;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.Configuration;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Options;
 using NUglify.Css;
-using NUglify.JavaScript;
 using Smidge.FileProcessors;
 using Smidge.Options;
 
@@ -39,7 +35,6 @@ namespace Smidge.Nuglify
 
         public static void UseSmidgeNuglify(this IApplicationBuilder app, bool useEndpointRouting = true)
         {
-#if NETCORE3_0
             //NOTE: It's no longer polite to just call UseMVC as it enables things that the developer may 
             //not need and the dev must disable EndpointRouting - so we let the dev decide.
             //with core 3.0 you have to explicitly disable EndpointRouting se we default to on here 
@@ -59,7 +54,6 @@ namespace Smidge.Nuglify
             }
             else
             {
-#endif
                 //Create custom route
                 app.UseMvc(routes =>
                 {
@@ -70,11 +64,7 @@ namespace Smidge.Nuglify
                         options.Value.UrlOptions.BundleFilePath + "/nmap/{bundle}",
                         new { controller = "NuglifySourceMap", action = "SourceMap" });
                 });
-#if NETCORE3_0
             }
-#endif
-
-            
 
         }
     }
