@@ -8,6 +8,7 @@
 
     @await SmidgeHelper.JsHereAsync("libs-js", debug: true)
     @await SmidgeHelper.CssHereAsync("libs-css", debug: true)
+    @*@await SmidgeHelper.CssHereAsync("less-test", debug: true)*@
 
     @{
         SmidgeHelper
@@ -21,14 +22,14 @@
             .CreateCssBundle("inline-css-bundle")
             .RequiresCss("~/Css/test1.css", "~/Css/test2.css");
 
-        SmidgeHelper.RequiresJs("~/Js/Folder*js");
-        SmidgeHelper.RequiresCss("~/Css/Folder*css");
+        SmidgeHelper.RequiresJs("~/Js/Folder/*.js");
+        SmidgeHelper.RequiresCss("~/Css/Folder/*.css");
     }
 
     <script src="inline-js-bundle" type="text/javascript" debug="true"></script>
-    <link rel="stylesheet" type="text/css" href="//cdnjs.cloudflare.com/ajax/libs/foundation/5.5.2/css/foundation.min.css"/>
-    <link rel="stylesheet" href="inline-css-bundle" debug="true"/>
-    <link rel="stylesheet" href="test-bundle-4" debug="true" />    
+    <link rel="stylesheet" type="text/css" href="//cdnjs.cloudflare.com/ajax/libs/foundation/5.5.2/css/foundation.min.css" />
+    <link rel="stylesheet" href="inline-css-bundle" debug="true" />
+    <link rel="stylesheet" href="test-bundle-4" debug="true" />
 </head>
 <body>
 
@@ -38,6 +39,9 @@
     @await SmidgeHelper.JsHereAsync(debug: true)
     @await SmidgeHelper.JsHereAsync("test-bundle-1", debug: true)
     @await SmidgeHelper.JsHereAsync("test-bundle-2", debug: true)
+    @await SmidgeHelper.JsHereAsync("test-bundle-10", debug: true)
+
+    <script src="~/Js/non-bundled.js" type="text/javascript"></script>
 
 </body>
 </html>
