@@ -7,6 +7,7 @@
 
     @await SmidgeHelper.JsHereAsync("libs-js")
     @await SmidgeHelper.CssHereAsync("libs-css")
+    @*@await SmidgeHelper.CssHereAsync("less-test")*@
 
     @{
         SmidgeHelper
@@ -23,16 +24,16 @@
         // This is here just to test what hapens when files don't exist
         SmidgeHelper
             .CreateJsBundle("no-files")
-            .RequiresJs("~/Js/not-found*js");
+            .RequiresJs("~/Js/not-found/*.js");
 
-        SmidgeHelper.RequiresJs("~/Js/Folder*js");
-        SmidgeHelper.RequiresCss("~/Css/Folder*css");
+        SmidgeHelper.RequiresJs("~/Js/Folder/*.js");
+        SmidgeHelper.RequiresCss("~/Css/Folder/*.css");
     }
 
     <script src="inline-js-bundle" type="text/javascript"></script>
     <link rel="stylesheet" type="text/css" href="//cdnjs.cloudflare.com/ajax/libs/foundation/5.5.2/css/foundation.min.css" />
-    <link rel="stylesheet" href="inline-css-bundle"/>
-    <link rel="stylesheet" href="test-bundle-4" />    
+    <link rel="stylesheet" href="inline-css-bundle" />
+    <link rel="stylesheet" href="test-bundle-4" />
 </head>
 <body>
 
@@ -43,6 +44,7 @@
     @await SmidgeHelper.JsHereAsync("test-bundle-1")
     @await SmidgeHelper.JsHereAsync("test-bundle-2")
     @await SmidgeHelper.JsHereAsync("no-files", debug: false)
+    @await SmidgeHelper.JsHereAsync("test-bundle-10")
 
     <script src="~/Js/non-bundled.js" type="text/javascript"></script>
 </body>
