@@ -20,8 +20,8 @@
             .CreateCssBundle("inline-css-bundle")
             .RequiresCss("~/Css/test1.css", "~/Css/test2.css");
 
-        SmidgeHelper.RequiresJs("~/Js/Folder*js");
-        SmidgeHelper.RequiresCss("~/Css/Folder*css");
+        SmidgeHelper.RequiresJs("~/Js/Folder/*.js");
+        SmidgeHelper.RequiresCss("~/Css/Folder/*.css");
     }
 
     <script src="inline-js-bundle" type="text/javascript"></script>
@@ -37,6 +37,8 @@
     @await SmidgeHelper.JsHereAsync(debug: false)
     @await SmidgeHelper.JsHereAsync("test-bundle-1")
     @await SmidgeHelper.JsHereAsync("test-bundle-2")
+
+    <script src="~/Js/non-bundled.js" type="text/javascript"></script>
 
 </body>
 </html>
