@@ -3,6 +3,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Smidge.Models;
 using System;
 using System.Linq;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc.Filters;
 using Smidge.Hashing;
 
@@ -13,18 +14,14 @@ namespace Smidge.Controllers
     /// </summary>
     public class CheckNotModifiedAttribute : Attribute, IFilterFactory, IOrderedFilter
     {
-        public int Order { get; set; }
-
         public IFilterMetadata CreateInstance(IServiceProvider serviceProvider)
         {
             return new CheckNotModifiedFilter(serviceProvider.GetRequiredService<IHasher>());
         }
 
-        /// <summary>
-        /// Gets a value that indicates if the result of <see cref="M:Microsoft.AspNetCore.Mvc.Filters.IFilterFactory.CreateInstance(System.IServiceProvider)" />
-        /// can be reused across requests.
-        /// </summary>
         public bool IsReusable => true;
+
+        public int Order { get; set; }
 
         public sealed class CheckNotModifiedFilter : IActionFilter
         {
@@ -37,14 +34,12 @@ namespace Smidge.Controllers
 
             public void OnActionExecuting(ActionExecutingContext context)
             {
-                if (!context.ActionArguments.Any()) return;
+                if (context.ActionArguments.Count == 0)
+                    return;
 
                 //put the model in the context, we'll resolve that after it's executed
-                var file = context.ActionArguments.First().Value as RequestModel;
-                if (file != null)
-                {
+                if (context.ActionArguments.First().Value is RequestModel file)
                     context.HttpContext.Items[nameof(CheckNotModifiedAttribute)] = file;
-                }
             }
 
             /// <summary>
@@ -56,28 +51,27 @@ namespace Smidge.Controllers
                 if (context.Exception != null) return;
 
                 //get the model from the items
-                if (!context.HttpContext.Items.ContainsKey(nameof(CheckNotModifiedAttribute))) return;
-                var file = context.HttpContext.Items[nameof(CheckNotModifiedAttribute)] as RequestModel;
-                if (file == null) return;
-
-                //Don't execute when the request is in Debug
-                if (file.Debug) return;
-
-                var etag = _hasher.Hash(file.FileKey + file.Compression + file.Mime);
-
-                var isDifferent = context.HttpContext.Request.HasETagBeenModified(etag);
-                var hasChanged = context.HttpContext.Request.HasRequestBeenModifiedSince(file.LastFileWriteTime.ToUniversalTime());
-                if (!isDifferent || !hasChanged)
+                if (context.HttpContext.Items.TryGetValue(nameof(CheckNotModifiedAttribute), out var requestModel) && requestModel is RequestModel file)
                 {
-                    ReturnNotModified(context);
+                    //Don't execute when the request is in Debug
+                    if (file.Debug)
+                        return;
+
+                    var etag = _hasher.Hash(file.FileKey + file.Compression + file.Mime);
+
+                    var isDifferent = context.HttpContext.Request.HasETagBeenModified(etag);
+                    var hasChanged = context.HttpContext.Request.HasRequestBeenModifiedSince(file.LastFileWriteTime.ToUniversalTime());
+                    if (!isDifferent || !hasChanged)
+                    {
+                        ReturnNotModified(context);
+                    }
                 }
             }
 
-            private void ReturnNotModified(ActionExecutedContext context)
+            private static void ReturnNotModified(ActionExecutedContext context)
             {
-                context.Result = new StatusCodeResult(HttpConstants.NotModified304);
+                context.Result = new StatusCodeResult(StatusCodes.Status304NotModified);
             }
         }
-
     }
 }
