@@ -10,6 +10,8 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.FileProviders;
 using Microsoft.Extensions.Logging;
 using Smidge.FileProcessors;
+using Smidge.Cache;
+using Microsoft.AspNetCore.Authorization;
 
 namespace Smidge.Controllers
 {
@@ -20,16 +22,17 @@ namespace Smidge.Controllers
     [AddCompressionHeader(Order = 0)]
     [AddExpiryHeaders(Order = 1)]
     [CheckNotModified(Order = 2)]
-    [CompositeFileCacheFilter(Order = 3)]        
+    [CompositeFileCacheFilter(Order = 3)]
+    [AllowAnonymous]
     public class SmidgeController : Controller
     {
-        private readonly FileSystemHelper _fileSystemHelper;
+        private readonly ISmidgeFileSystem _fileSystem;
         private readonly IBundleManager _bundleManager;
         private readonly IBundleFileSetGenerator _fileSetGenerator;
         private readonly PreProcessPipelineFactory _processorFactory;
-        private readonly PreProcessManager _preProcessManager;
+        private readonly IPreProcessManager _preProcessManager;
         private readonly ILogger _logger;
-        
+
         /// <summary>
         /// Constructor
         /// </summary>
@@ -40,23 +43,18 @@ namespace Smidge.Controllers
         /// <param name="preProcessManager"></param>
         /// <param name="logger"></param>
         public SmidgeController(
-            FileSystemHelper fileSystemHelper, 
+            ISmidgeFileSystem fileSystemHelper,
             IBundleManager bundleManager,
             IBundleFileSetGenerator fileSetGenerator,
             PreProcessPipelineFactory processorFactory,
-            PreProcessManager preProcessManager,
+            IPreProcessManager preProcessManager,
             ILogger<SmidgeController> logger)
         {
-            if (fileSystemHelper == null) throw new ArgumentNullException(nameof(fileSystemHelper));
-            if (bundleManager == null) throw new ArgumentNullException(nameof(bundleManager));
-            if (fileSetGenerator == null) throw new ArgumentNullException(nameof(fileSetGenerator));
-            if (processorFactory == null) throw new ArgumentNullException(nameof(processorFactory));
-            if (preProcessManager == null) throw new ArgumentNullException(nameof(preProcessManager));
-            _fileSystemHelper = fileSystemHelper;
-            _bundleManager = bundleManager;
-            _fileSetGenerator = fileSetGenerator;
-            _processorFactory = processorFactory;
-            _preProcessManager = preProcessManager;
+            _fileSystem = fileSystemHelper ?? throw new ArgumentNullException(nameof(fileSystemHelper));
+            _bundleManager = bundleManager ?? throw new ArgumentNullException(nameof(bundleManager));
+            _fileSetGenerator = fileSetGenerator ?? throw new ArgumentNullException(nameof(fileSetGenerator));
+            _processorFactory = processorFactory ?? throw new ArgumentNullException(nameof(processorFactory));
+            _preProcessManager = preProcessManager ?? throw new ArgumentNullException(nameof(preProcessManager));
             _logger = logger;
         }
 
@@ -66,7 +64,7 @@ namespace Smidge.Controllers
         /// <param name="bundleModel">The bundle model</param>
         /// <returns></returns>       
         public async Task<IActionResult> Bundle(
-            [FromServices]BundleRequestModel bundleModel)
+            [FromServices] BundleRequestModel bundleModel)
         {
             if (!_bundleManager.TryGetValue(bundleModel.FileKey, out Bundle foundBundle))
             {
@@ -74,15 +72,26 @@ namespace Smidge.Controllers
             }
 
             var bundleOptions = foundBundle.GetBundleOptions(_bundleManager, bundleModel.Debug);
-            
+
+            var cacheBusterValue = bundleModel.ParsedPath.CacheBusterValue;
+
             //now we need to determine if this bundle has already been created
-            var compositeFilePath = new FileInfo(_fileSystemHelper.GetCurrentCompositeFilePath(bundleModel.CacheBuster, bundleModel.Compression, bundleModel.FileKey));
-            if (compositeFilePath.Exists)
+            var cacheFile = _fileSystem.CacheFileSystem.GetCachedCompositeFile(cacheBusterValue, bundleModel.Compression, bundleModel.FileKey, out var cacheFilePath);
+            if (cacheFile.Exists)
             {
                 _logger.LogDebug($"Returning bundle '{bundleModel.FileKey}' from cache");
 
-                //this is already processed, return it
-                return File(compositeFilePath.OpenRead(), bundleModel.Mime);
+
+                if (!string.IsNullOrWhiteSpace(cacheFile.PhysicalPath))
+                {
+                    //if physical path is available then it's the physical file system, in which case we'll deliver the file with the PhysicalFileResult
+                    //FilePathResult uses IHttpSendFileFeature which is a native host option for sending static files                    
+                    return PhysicalFile(cacheFile.PhysicalPath, bundleModel.Mime);
+                }
+                else
+                {
+                    return File(cacheFile.CreateReadStream(), bundleModel.Mime);
+                }
             }
 
             //the bundle doesn't exist so we'll go get the files, process them and create the bundle
@@ -99,8 +108,8 @@ namespace Smidge.Controllers
             {
                 return NotFound();
             }
-            
-            using (var bundleContext = new BundleContext(bundleModel, compositeFilePath))
+
+            using (var bundleContext = new BundleContext(cacheBusterValue, bundleModel, cacheFilePath))
             {
                 var watch = new Stopwatch();
                 watch.Start();
@@ -113,22 +122,26 @@ namespace Smidge.Controllers
                 }
 
                 //Get each file path to it's hashed location since that is what the pre-processed file will be saved as
-                Lazy<IFileInfo> fi;
-                var filePaths = files.Select(
-                    x => _fileSystemHelper.GetCacheFilePath(x, bundleOptions.FileWatchOptions.Enabled, Path.GetExtension(x.FilePath), bundleModel.CacheBuster, out fi));
+                var fileInfos = files.Select(x => _fileSystem.CacheFileSystem.GetCacheFile(
+                    x,
+                    () => _fileSystem.GetRequiredFileInfo(x),
+                    bundleOptions.FileWatchOptions.Enabled,
+                    Path.GetExtension(x.FilePath),
+                    cacheBusterValue,
+                    out _));
 
-                using (var resultStream = await GetCombinedStreamAsync(filePaths, bundleContext))
+                using (var resultStream = await GetCombinedStreamAsync(fileInfos, bundleContext))
                 {
                     //compress the response (if enabled)
                     var compressedStream = await Compressor.CompressAsync(
                         //do not compress anything if it's not enabled in the bundle options
-                        bundleOptions.CompressResult ? bundleModel.Compression : CompressionType.none,
+                        bundleOptions.CompressResult ? bundleModel.Compression : CompressionType.None,
                         resultStream);
 
                     //save the resulting compressed file, if compression is not enabled it will just save the non compressed format
                     // this persisted file will be used in the CheckNotModifiedAttribute which will short circuit the request and return
                     // the raw file if it exists for further requests to this path
-                    await CacheCompositeFileAsync(compositeFilePath, compressedStream);
+                    await CacheCompositeFileAsync(_fileSystem.CacheFileSystem, cacheFilePath, compressedStream);
 
                     _logger.LogDebug($"Processed bundle '{bundleModel.FileKey}' in {watch.ElapsedMilliseconds}ms");
 
@@ -144,68 +157,72 @@ namespace Smidge.Controllers
         /// <param name="file"></param>
         /// <returns></returns>
         public async Task<IActionResult> Composite(
-             [FromServices]CompositeFileModel file)
+             [FromServices] CompositeFileModel file)
         {
             if (!file.ParsedPath.Names.Any())
             {
                 return NotFound();
             }
 
-            var compositeFilePath = new FileInfo(_fileSystemHelper.GetCurrentCompositeFilePath(file.CacheBuster, file.Compression, file.FileKey));
+            var defaultBundleOptions = _bundleManager.GetDefaultBundleOptions(false);
+            var cacheBusterValue = file.ParsedPath.CacheBusterValue;
 
-            if (compositeFilePath.Exists)
+            var cacheFile = _fileSystem.CacheFileSystem.GetCachedCompositeFile(cacheBusterValue, file.Compression, file.FileKey, out var cacheFilePath);
+
+            if (cacheFile.Exists)
             {
                 //this is already processed, return it
-                return File(compositeFilePath.OpenRead(), file.Mime);
+                if (!string.IsNullOrWhiteSpace(cacheFile.PhysicalPath))
+                {
+                    //if physical path is available then it's the physical file system, in which case we'll deliver the file with the PhysicalFileResult
+                    //FilePathResult uses IHttpSendFileFeature which is a native host option for sending static files                    
+                    return PhysicalFile(cacheFile.PhysicalPath, file.Mime);
+                }
+                else
+                {
+                    return File(cacheFile.CreateReadStream(), file.Mime);
+                }
             }
-            
-            using (var bundleContext = new BundleContext(file, compositeFilePath))
-            {
-                var filePaths = file.ParsedPath.Names.Select(filePath =>
-                    Path.Combine(
-                        _fileSystemHelper.CurrentCacheFolder,
-                        file.ParsedPath.Version,
-                        filePath + file.Extension));
 
-                using (var resultStream = await GetCombinedStreamAsync(filePaths, bundleContext))
+            using (var bundleContext = new BundleContext(cacheBusterValue, file, cacheFilePath))
+            {
+                var files = file.ParsedPath.Names.Select(filePath =>
+                    _fileSystem.CacheFileSystem.GetRequiredFileInfo(
+                        $"{file.ParsedPath.CacheBusterValue}/{filePath + file.Extension}"));
+
+                using (var resultStream = await GetCombinedStreamAsync(files, bundleContext))
                 {
                     var compressedStream = await Compressor.CompressAsync(file.Compression, resultStream);
 
-                    await CacheCompositeFileAsync(compositeFilePath, compressedStream);
+                    await CacheCompositeFileAsync(_fileSystem.CacheFileSystem, cacheFilePath, compressedStream);
 
                     return File(compressedStream, file.Mime);
                 }
             }
         }
 
-        private static async Task CacheCompositeFileAsync(FileInfo compositeFilePath, Stream compositeStream)
+        private static async Task CacheCompositeFileAsync(ICacheFileSystem cacheProvider, string filePath, Stream compositeStream)
         {
-            //ensure it exists
-            compositeFilePath.Directory.Create();            
-            compositeStream.Position = 0;            
-            using (var fs = compositeFilePath.Create())
-            {
-                await compositeStream.CopyToAsync(fs);
-            }
-            compositeStream.Position = 0;
+            await cacheProvider.WriteFileAsync(filePath, compositeStream);
+            if (compositeStream.CanSeek)
+                compositeStream.Position = 0;
         }
 
         /// <summary>
         /// Combines files into a single stream
         /// </summary>
-        /// <param name="filePaths"></param>
+        /// <param name="files"></param>
         /// <param name="bundleContext"></param>
         /// <returns></returns>
-        private async Task<Stream> GetCombinedStreamAsync(IEnumerable<string> filePaths, BundleContext bundleContext)
+        private async Task<Stream> GetCombinedStreamAsync(IEnumerable<IFileInfo> files, BundleContext bundleContext)
         {
             //TODO: Here we need to be able to prepend/append based on a "BundleContext" (or similar)
 
             List<Stream> inputs = null;
             try
             {
-                inputs = filePaths.Where(System.IO.File.Exists)
-                    .Select(System.IO.File.OpenRead)
-                    .Cast<Stream>()
+                inputs = files.Where(x => x.Exists)
+                    .Select(x => x.CreateReadStream())
                     .ToList();
 
                 var delimeter = bundleContext.BundleRequest.Extension == ".js" ? ";\n" : "\n";
@@ -223,9 +240,5 @@ namespace Smidge.Controllers
                 }
             }
         }
-
-        
     }
-
-
 }
