@@ -2,7 +2,6 @@
 using System;
 using System.Linq;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
-using Smidge.Cache;
 
 namespace Smidge.Models
 {
@@ -12,7 +11,7 @@ namespace Smidge.Models
     /// </summary>
     public class BundleRequestModel : RequestModel
     {
-        public BundleRequestModel(IUrlManager urlManager, IActionContextAccessor accessor, IRequestHelper requestHelper, IBundleManager bundleManager, CacheBusterResolver cacheBusterResolver)
+        public BundleRequestModel(IUrlManager urlManager, IActionContextAccessor accessor, IRequestHelper requestHelper, IBundleManager bundleManager)
             : base("bundle", urlManager, accessor, requestHelper)
         {
             //TODO: Pretty sure if we want to control the caching of the file, we'll have to retrieve the bundle definition here
@@ -27,17 +26,13 @@ namespace Smidge.Models
 
             FileKey = ParsedPath.Names.Single();
 
-            Bundle bundle;
-            if (!bundleManager.TryGetValue(FileKey, out bundle))
+            if (!bundleManager.TryGetValue(FileKey, out Bundle bundle))
             {
                 throw new InvalidOperationException("No bundle found with key " + FileKey);
             }
             Bundle = bundle;
-
-            CacheBuster = cacheBusterResolver.GetCacheBuster(bundle.GetBundleOptions(bundleManager, Debug).GetCacheBusterType());                
         }
 
-        public override ICacheBuster CacheBuster { get; }
         public Bundle Bundle { get; }        
         public override string FileKey { get; }
     }
