@@ -1,6 +1,5 @@
 ﻿using Smidge.CompositeFiles;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
-using Smidge.Cache;
 using Smidge.Hashing;
 
 namespace Smidge.Models
@@ -8,16 +7,13 @@ namespace Smidge.Models
     public class CompositeFileModel : RequestModel
     {
 
-        public CompositeFileModel(IHasher hasher, IUrlManager urlManager, IActionContextAccessor accessor, IRequestHelper requestHelper, IBundleManager bundleManager, CacheBusterResolver cacheBusterResolver)
+        public CompositeFileModel(IHasher hasher, IUrlManager urlManager, IActionContextAccessor accessor, IRequestHelper requestHelper)
             : base("file", urlManager, accessor, requestHelper)
         {
             //Creates a single hash of the full url (which can include many files)
             FileKey = hasher.Hash(string.Join(".", ParsedPath.Names));
-
-            CacheBuster = cacheBusterResolver.GetCacheBuster(bundleManager.GetDefaultBundleOptions(false).GetCacheBusterType());
         }
 
-        public override ICacheBuster CacheBuster { get; }
         public override string FileKey { get; }
     }
 }
