@@ -1,7 +1,10 @@
 ﻿using System;
-using System.Diagnostics;
+using System.Collections.Generic;
 using System.Text.RegularExpressions;
 using Microsoft.AspNetCore.Http;
+using Microsoft.Extensions.Primitives;
+using Microsoft.Net.Http.Headers;
+using Smidge.Models;
 
 namespace Smidge
 {
@@ -10,9 +13,8 @@ namespace Smidge
         private readonly IWebsiteInfo _siteInfo;
 
         public RequestHelper(IWebsiteInfo siteInfo)
-        {            
-            if (siteInfo == null) throw new ArgumentNullException(nameof(siteInfo));
-            _siteInfo = siteInfo;
+        {
+            _siteInfo = siteInfo ?? throw new ArgumentNullException(nameof(siteInfo));
         }
 
         public bool IsExternalRequestPath(string path)
@@ -24,6 +26,32 @@ namespace Smidge
                 return true;
             }
             return false;
+        }
+
+        public string Content(IWebFile file)
+        {
+            if (string.IsNullOrEmpty(file.FilePath))
+                return null;
+
+            //if this is a protocol-relative/protocol-less uri, then we need to add the protocol for the remaining
+            // logic to work properly
+            if (file.FilePath.StartsWith("//"))
+            {
+                var scheme = _siteInfo.GetBaseUrl().Scheme;
+                return Regex.Replace(file.FilePath, @"^\/\/", string.Format("{0}{1}", scheme, SmidgeConstants.SchemeDelimiter));
+            }
+
+            var filePath = Content(file.FilePath);
+            if (filePath == null) return null;
+
+            var requestPath = file.RequestPath != null ? Content(file.RequestPath) : string.Empty;
+
+            if (requestPath.EndsWith('/'))
+            {
+                requestPath.TrimEnd('/');
+            }
+
+            return string.Concat(requestPath, filePath);
         }
 
         /// <summary>
@@ -41,7 +69,7 @@ namespace Smidge
             if (path.StartsWith("//"))
             {
                 var scheme = _siteInfo.GetBaseUrl().Scheme;
-                return Regex.Replace(path, @"^\/\/", string.Format("{0}{1}", scheme, Constants.SchemeDelimiter));
+                return Regex.Replace(path, @"^\/\/", string.Format("{0}{1}", scheme, SmidgeConstants.SchemeDelimiter));
             }
 
             //This code is taken from the UrlHelper code ... which shouldn't need to be tucked away in there
@@ -62,33 +90,45 @@ namespace Smidge
         /// If IE 6 is detected, we will ignore compression as it's known that some versions of IE 6
         /// have issues with it.
         /// </summary>
-        public CompressionType GetClientCompression(IHeaderDictionary headers)
+        public CompressionType GetClientCompression(IDictionary<string, StringValues> headers)
         {
-            var type = CompressionType.none;
-            var agentHeader = (string)headers[HttpConstants.UserAgent];
+            var type = CompressionType.None;
+            var agentHeader = (string)headers[HeaderNames.UserAgent];
             if (agentHeader != null && agentHeader.Contains("MSIE 6"))
             {
                 return type;
             }
 
-            string acceptEncoding = headers[HttpConstants.AcceptEncoding];
+            string acceptEncoding = headers[HeaderNames.AcceptEncoding];
 
             if (!string.IsNullOrEmpty(acceptEncoding))
             {
-                string[] supported = acceptEncoding.Split(',');
-                //get the first type that we support
-                for (var i = 0; i < supported.Length; i++)
+                string[] acceptedEncodings = acceptEncoding.Split(',');
+
+                // Prefer in order: Brotli, GZip, Deflate.
+                // https://www.iana.org/assignments/http-parameters/http-parameters.xml#http-content-coding-registry
+                for (var i = 0; i < acceptedEncodings.Length; i++)
                 {
-                    if (supported[i].Contains("deflate"))
+                    var encoding = acceptedEncodings[i].Trim();
+
+                    CompressionType parsed = CompressionType.Parse(encoding);
+
+                    // Brotli is typically last in the accept encoding header.
+                    if (parsed == CompressionType.Brotli)
                     {
-                        type = CompressionType.deflate;
-                        break;
+                        return CompressionType.Brotli;
                     }
-                    else if (supported[i].Contains("gzip")) //sometimes it could be x-gzip!
+                    
+                    // Not pack200-gzip.
+                    if (parsed == CompressionType.GZip)
                     {
-                        type = CompressionType.gzip;
-                        break;
+                        type = CompressionType.GZip;
                     }
+                        
+                    if (type != CompressionType.GZip && parsed == CompressionType.Deflate)
+                    {
+                        type = CompressionType.Deflate;
+                    }   
                 }
             }
 
