@@ -1,85 +1,79 @@
 ﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc.ApplicationParts;
 using Microsoft.AspNetCore.Mvc.Infrastructure;
-using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.FileProviders;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Options;
 using Smidge.Cache;
 using Smidge.CompositeFiles;
-using Smidge.Controllers;
 using Smidge.FileProcessors;
 using Smidge.Hashing;
 using Smidge.Models;
 using Smidge.Options;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Runtime.CompilerServices;
 
 [assembly: InternalsVisibleTo("Smidge.Tests")]
 
 namespace Smidge
 {
+
     public static class SmidgeStartup
     {
         //For .net core 3.0, call this before AddControllers/AddControllersWithViews etc
         //If you call it after, then if AddControllersAsServices was used before you need
         //to tell smidge to do so
 
-#if NETCORE3_0
-        public static IServiceCollection AddSmidge(this IServiceCollection services,
-            IConfiguration smidgeConfiguration = null,
-            IFileProvider fileProvider = null)
-#else
-
-        public static IServiceCollection AddSmidge(this IServiceCollection services,
-            IConfiguration smidgeConfiguration = null,
-            IFileProvider fileProvider = null)
-#endif
+        public static IServiceCollection AddSmidge(this IServiceCollection services, IConfiguration smidgeConfiguration = null)
         {
             services.TryAddSingleton<IHttpContextAccessor, HttpContextAccessor>();
             services.TryAddSingleton<IActionContextAccessor, ActionContextAccessor>();
 
             services.AddTransient<IConfigureOptions<SmidgeOptions>, SmidgeOptionsSetup>();
 
-            services.AddSingleton<PreProcessManager>();
+            services.AddSingleton<IPreProcessManager, PreProcessManager>();
             services.AddSingleton<IRequestHelper, RequestHelper>();
             services.AddSingleton<IWebsiteInfo, AutoWebsiteInfo>();
             services.AddSingleton<IBundleFileSetGenerator, BundleFileSetGenerator>();
             services.AddSingleton<IHasher, Crc32Hasher>();
+            services.AddSingleton<IFileProviderFilter, DefaultFileProviderFilter>();
             services.AddSingleton<IBundleManager, BundleManager>();
             services.AddSingleton<PreProcessPipelineFactory>();
-            services.AddSingleton<FileSystemHelper>(p =>
+            services.AddSingleton<ISmidgeFileSystem>(p =>
             {
-#if NETCORE3_0                     
                 var hosting = p.GetRequiredService<IWebHostEnvironment>();
-#else
-                var hosting = p.GetRequiredService<IHostingEnvironment>();
-#endif
-                var provider = fileProvider ?? hosting.WebRootFileProvider;
 
-                return new FileSystemHelper(hosting, p.GetRequiredService<ISmidgeConfig>(), provider, p.GetRequiredService<IHasher>(), p.GetRequiredService<IWebsiteInfo>());
+                //resolve the ISmidgeFileProvider if there is one
+                var provider = p.GetService<ISmidgeFileProvider>() ?? hosting.WebRootFileProvider;
+                return new SmidgeFileSystem(
+                    provider,
+                    p.GetRequiredService<IFileProviderFilter>(),
+                    p.GetRequiredService<ICacheFileSystem>(),
+                    p.GetRequiredService<IWebsiteInfo>());
             });
+            
+            services.AddSingleton<ICacheFileSystem>(p => PhysicalFileCacheFileSystem.CreatePhysicalFileCacheFileSystem(
+                p.GetRequiredService<IHasher>(),
+                p.GetRequiredService<ISmidgeConfig>(),
+                p.GetRequiredService<IHostEnvironment>()));
+
             services.AddSingleton<ISmidgeConfig>((p) =>
             {
                 if (smidgeConfiguration == null)
                 {
-#if NETCORE3_0                     
-                    return new SmidgeConfig(p.GetRequiredService<IWebHostEnvironment>());
-#else
-                    return new SmidgeConfig(p.GetRequiredService<IHostingEnvironment>());
-#endif
+                    return new SmidgeConfig();
                 }
                 return new SmidgeConfig(smidgeConfiguration);
             });
 
             services.AddSingleton<ICacheBuster, ConfigCacheBuster>();
             services.AddSingleton<ICacheBuster, AppDomainLifetimeCacheBuster>();
+            services.AddSingleton<ICacheBuster, TimestampCacheBuster>();
             services.AddSingleton<CacheBusterResolver>();
 
             //These all execute as part of the request/scope            
@@ -113,15 +107,11 @@ namespace Smidge
 
 
 
-#if NETCORE3_0
         public static void UseSmidge(this IApplicationBuilder app, Action<IBundleManager> configureBundles = null, bool useEndpointRouting = true)
-#else
-        public static void UseSmidge(this IApplicationBuilder app, Action<IBundleManager> configureBundles = null)
-#endif
         {
             //Creates custom routes 
             var options = app.ApplicationServices.GetRequiredService<IOptions<SmidgeOptions>>();
-#if NETCORE3_0
+
             //NOTE: It's no longer polite to just call UseMVC as it enables things that the developer may 
             //not need and the dev must disable EndpointRouting - so we let the dev decide.
             //with core 3.0 you have to explicitly disable EndpointRouting se we default to on here 
@@ -142,7 +132,7 @@ namespace Smidge
             }
             else
             {
-#endif
+
                 app.UseMvc(routes =>
                 {
                     routes.MapRoute(
@@ -155,66 +145,64 @@ namespace Smidge
                         options.Value.UrlOptions.BundleFilePath + "/{bundle}",
                         new { controller = "Smidge", action = "Bundle" });
                 });
-#if NETCORE3_0
             }
-#endif
+
             if (configureBundles != null)
             {
                 var bundleManager = app.ApplicationServices.GetRequiredService<IBundleManager>();
                 configureBundles(bundleManager);
 
                 var cacheBusterResolver = app.ApplicationServices.GetRequiredService<CacheBusterResolver>();
+                var fileSystem = app.ApplicationServices.GetRequiredService<ISmidgeFileSystem>();
 
                 //TODO: Now that they are configured we need to wire up the file watching event handlers
                 // to the bundle manager, currently these are on the Bundle, but that is not good enough
                 // since we need the bundle name
-                foreach (var webFileType in new[] {WebFileType.Css, WebFileType.Js })
+                foreach (var webFileType in new[] { WebFileType.Css, WebFileType.Js })
                 {
-                    var names = bundleManager.GetBundleNames(webFileType);
-                    foreach (var cssName in names)
+                    var bundles = bundleManager.GetBundles(webFileType);
+                    foreach (var bundle in bundles)
                     {
-                        var bundle = bundleManager.GetBundle(cssName);
-                        WireUpFileWatchEventHandlers(bundleManager, cacheBusterResolver, cssName, bundle);
+                        WireUpFileWatchEventHandlers(cacheBusterResolver, fileSystem, bundle);
                     }
                 }
-            }    
+            }
         }
 
-        private static void WireUpFileWatchEventHandlers(IBundleManager bundleManager, CacheBusterResolver cacheBusterResolver, string bundleName, Bundle bundle)
+        private static void WireUpFileWatchEventHandlers(CacheBusterResolver cacheBusterResolver, ISmidgeFileSystem fileSystem, Bundle bundle)
         {
             if (bundle.BundleOptions == null) return;
 
             if (bundle.BundleOptions.DebugOptions.FileWatchOptions.Enabled)
             {
-                bundle.BundleOptions.DebugOptions.FileWatchOptions.FileModified += FileWatchOptions_FileModified(bundleManager, cacheBusterResolver, bundleName, bundle, true);
+                bundle.BundleOptions.DebugOptions.FileWatchOptions.FileModified += FileWatchOptions_FileModified(cacheBusterResolver, fileSystem, bundle.Name);
             }
             if (bundle.BundleOptions.ProductionOptions.FileWatchOptions.Enabled)
             {
-                bundle.BundleOptions.ProductionOptions.FileWatchOptions.FileModified += FileWatchOptions_FileModified(bundleManager, cacheBusterResolver, bundleName, bundle, false);
+                bundle.BundleOptions.ProductionOptions.FileWatchOptions.FileModified += FileWatchOptions_FileModified(cacheBusterResolver, fileSystem, bundle.Name);
             }
         }
 
-        private static EventHandler<FileWatchEventArgs> FileWatchOptions_FileModified(IBundleManager bundleManager, CacheBusterResolver cacheBusterResolver, string bundleName, Bundle bundle, bool debug)
+        private static EventHandler<FileWatchEventArgs> FileWatchOptions_FileModified(CacheBusterResolver cacheBusterResolver, ISmidgeFileSystem fileSystem, string bundleName)
         {
             return (sender, args) =>
             {
-                FileWatchOptions_FileModified(bundleManager, cacheBusterResolver, bundleName, bundle, debug, args);
+                FileWatchOptions_FileModified(cacheBusterResolver, fileSystem, bundleName, args);
             };
         }
 
-        private static void FileWatchOptions_FileModified(IBundleManager bundleManager, CacheBusterResolver cacheBusterResolver, string bundleName, Bundle bundle, bool debug, FileWatchEventArgs e)
+        //async void = ok here sincew this is an event handler
+        private static async void FileWatchOptions_FileModified(CacheBusterResolver cacheBusterResolver, ISmidgeFileSystem fileSystem, string bundleName, FileWatchEventArgs e)
         {
-            var cacheBuster = cacheBusterResolver.GetCacheBuster(bundle.GetBundleOptions(bundleManager, debug).GetCacheBusterType());                
+            var bundleOptions = e.File.BundleOptions;
+            var cacheBuster = cacheBusterResolver.GetCacheBuster(bundleOptions.GetCacheBusterType());
+            var cacheBusterValue = cacheBuster.GetValue();
 
-            //this file is part of this bundle, so the persisted processed/combined/compressed  will need to be 
+            //this file is part of this bundle, so the persisted processed/combined/compressed will need to be 
             // invalidated/deleted/renamed
-            foreach (var compressionType in new[] { CompressionType.deflate, CompressionType.gzip, CompressionType.none })
+            foreach (var compressionType in CompressionType.All)
             {
-                var compFilePath = e.FileSystemHelper.GetCurrentCompositeFilePath(cacheBuster, compressionType, bundleName);
-                if (File.Exists(compFilePath))
-                {
-                    File.Delete(compFilePath);
-                }
+                await fileSystem.CacheFileSystem.ClearCachedCompositeFileAsync(cacheBusterValue, compressionType, bundleName);
             }
         }
     }
