@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Moq;
@@ -20,15 +20,15 @@ namespace Smidge.Tests
 {
     public class SmidgeHelperTests
     {
-        private readonly ISmidgeConfig _config = Mock.Of<ISmidgeConfig>();
         private readonly IUrlManager _urlManager = Mock.Of<IUrlManager>();
-        private readonly IHostingEnvironment _hostingEnvironment = Mock.Of<IHostingEnvironment>();
         private readonly IFileProvider _fileProvider = Mock.Of<IFileProvider>();
+        private readonly ICacheFileSystem _cacheProvider = Mock.Of<ICacheFileSystem>();
+        private readonly IFileProviderFilter _fileProviderFilter = Mock.Of<IFileProviderFilter>();
         private readonly IHasher _hasher = Mock.Of<IHasher>();
         private readonly IEnumerable<IPreProcessor> _preProcessors = new List<IPreProcessor>();
         private readonly IBundleFileSetGenerator _fileSetGenerator;
         private readonly DynamicallyRegisteredWebFiles _dynamicallyRegisteredWebFiles;
-        private readonly FileSystemHelper _fileSystemHelper;
+        private readonly SmidgeFileSystem _fileSystemHelper;
         private readonly PreProcessManager _preProcessManager;
         private Mock<IOptions<SmidgeOptions>> _smidgeOptions;
         private readonly PreProcessPipelineFactory _processorFactory;
@@ -46,8 +46,7 @@ namespace Smidge.Tests
             _httpContextAccessor.Setup(x => x.HttpContext).Returns(_httpContext.Object);
 
             _dynamicallyRegisteredWebFiles = new DynamicallyRegisteredWebFiles();
-            _fileSystemHelper = new FileSystemHelper(_hostingEnvironment, _config, _fileProvider, _hasher, Mock.Of<IWebsiteInfo>());
-                        
+            _fileSystemHelper = new SmidgeFileSystem(_fileProvider, _fileProviderFilter, _cacheProvider, Mock.Of<IWebsiteInfo>());
             _smidgeOptions = new Mock<IOptions<SmidgeOptions>>();
             _smidgeOptions.Setup(opt => opt.Value).Returns(new SmidgeOptions
             {
@@ -59,9 +58,10 @@ namespace Smidge.Tests
             _bundleManager = new BundleManager(_smidgeOptions.Object, Mock.Of<ILogger<BundleManager>>());
             _preProcessManager = new PreProcessManager(
                 _fileSystemHelper,                
-                new CacheBusterResolver(Enumerable.Empty<ICacheBuster>()),
-                _bundleManager, Mock.Of<ILogger<PreProcessManager>>());
-            _fileSetGenerator = new BundleFileSetGenerator(_fileSystemHelper, _requestHelper, 
+                _bundleManager, 
+                Mock.Of<ILogger<PreProcessManager>>());
+            _fileSetGenerator = new BundleFileSetGenerator(
+                _fileSystemHelper, 
                 new FileProcessingConventions(_smidgeOptions.Object, new List<IFileProcessingConvention>()));
         }
 
@@ -70,7 +70,7 @@ namespace Smidge.Tests
         {
             var sut = new SmidgeHelper(
                 _fileSetGenerator,
-                _dynamicallyRegisteredWebFiles, _preProcessManager, _fileSystemHelper,
+                _dynamicallyRegisteredWebFiles, _preProcessManager, _fileSystemHelper, 
                 _hasher, _bundleManager, _processorFactory, _urlManager, _requestHelper,
                 _httpContextAccessor.Object,
                 new CacheBusterResolver(Enumerable.Empty<ICacheBuster>()));
